+++ parsers/mallox.py
@@ -1,7 +1,17 @@
+"""
++------------------------------+------------------+----------+
+| Description | Published Date | Victim's Website | Post URL |
++------------------------------+------------------+----------+
+|      X      |        X       |                  |     X    |
++------------------------------+------------------+----------+
+"""
+
 import os
 from bs4 import BeautifulSoup
 from sharedutils import errlog
 from parse import appender
+from datetime import datetime
+
 
 def main():
     for filename in os.listdir('source'):
@@ -16,7 +26,15 @@ def main():
                     description = ''
                     for p in div.find_all('p'):
                         description+=p.text + ' '
-                    appender(title, 'mallox', description)
+                    post = div.find('a', {'class': 'btn btn-primary btn-sm'})
+                    post = post.get('href')
+                    parts = filename.split('-')
+                    url = parts[1].replace('.html','')
+                    url = 'http://' + url + '.onion' + post
+                    publish = div.find('span', {'class': 'badge badge-info'}).text.strip()
+                    date_obj = datetime.strptime(publish, "%d/%m/%Y %H:%M")
+                    formatted_date = date_obj.strftime("%Y-%m-%d %H:%M:%S.%f")
+                    appender(title, 'mallox', description,"",formatted_date,url)
                 file.close()
         except:
             errlog('mallox: ' + 'parsing fail')