+++ parsers/vicesociety.py
@@ -1,3 +1,11 @@
+"""
++------------------------------+------------------+----------+
+| Description | Published Date | Victim's Website | Post URL |
++------------------------------+------------------+----------+
+|      X      |                |        X         |     X    |
++------------------------------+------------------+----------+
+"""
+
 import os
 from bs4 import BeautifulSoup
 from sharedutils import errlog
@@ -14,12 +22,14 @@ def main():
                 for div in divs_name:
                     try:
                         title = div.find("font", {"size":4}).text.strip() 
+                        post = div.find('a', {'style': 'text-decoration: none;'})      
+                        post = post.get('href')
                         for description in div.find_all("font", {"size":2, "color":"#5B61F6"}):
                             if description.b.text.strip().startswith("http"):
                                 website = description.get_text()
                             if not description.b.text.strip().startswith("http"):
                                 desc = description.get_text()
-                                appender(title, 'vicesociety', desc,website)
+                                appender(title, 'vicesociety', desc,website,post)
                     except:
                         pass
                 file.close()